Refactor to allow cascading methods in controllers and stop some of the function-object mutation
Removed some function-object mutation by using WeakMaps.  This ought to make things more efficient.  It would be ideal to mutate this throughout the bolt platform.

Added a cascade property lookup on controller methods.  Assumed to be false but if present and true, method paths with exactly the same path do not overwrite each other but cascade. Only the last cascade value is checked and this overrides for the whole stack.