Expand the controller scope
Complete the proxy for controller scope with all trap methods.  Stop expanding or tweaking with these after load by applying a deep freeze on the controllers.  This is a security method.  Ought to be enhance speed too as objects are constant.

Added a new deepFreeze method for recursive freezing of objects.

#35 #32