Add more inline docs and refactor string replacement code
Added some more inline docs surrounding database connections.  Refactored the mysql interface to make it clearer and to follow style used throughout bolt.

Refactoring of string replacements so in the common scenario of doing a series of regular expression replacements on a string one can now use a new function.  It will even create a bespoke function for you, which always runs a specific replacement sequence.

Ref #7, #22