Fix socket.io middleware and session data
Middleware in socket.io needed to run per packet not just per socket - fixed.  There was a balance between the socket.io request object and the request object of the actual socket.  Sometime the underlaying request needs changing as-when logging-in.  Used a proxy object to try and get this balance right.