Refactor socket.io routing and use again
This finally fixes the major problems with authentication.  We need to login/logout on the original request but do everything else on the socket.io request object.  Refactored middleware to fire on both socket routes and normal ones.  Refactored the socket.io request object creation so auth methods are bound to the original request rather than socket one.  Refactored other methods to bind to created object.