Add new injectors and handle view called components better
Added viaView and viaViewPath to component so we can tell if it is called in a component. Refactor values to handle component paths+query.  Added a display function to the injectors to handle the four cases of: View called component wanting json, view called component wanting html, controller wantng html and controller wanting json.