Refactor the async loading to make more use of the retry feature in require-extra.import()
Add the retry option to bolt module loading so there is now little need for bolt.ready().  Also, added new mapAsync method to bolt; it is similar to bluebird mapSeries.  This series mapping was need in the component loader which had an awful hack to load it twice thereby making the load order correct on the 2nd pass - this is now fixed.