Fixes for gulp tasks
Fixed the redux gulp task that just failed.  This was due to previous upgrades not worked into redux task.  Added removed code back into react task that was needed. Added cycling to gulp task adder so if a dep is not found it might be on next cycle. Upgraded replaceCSP to do cycles until no more replaces and deep gets.