Fixed the caching crash
Bundling in rollup was crashing when cache was being used.  This update fixes it.  To be 100% I am not sure what changed to break this or how it is now fixed.  I assume it is one of the deps either rollup or @simpo/rollup-vinyl-adaptor.  Since, we control the latter and not much has changed with it, it must be rollup.